refactor(runtime): remove redundant service_protocol.py re-export
Delete src/vibe3/runtime/service_protocol.py and redirect all consumers
to import ServiceBase directly from its canonical source vibe3.orchestra.domain_types.

This eliminates a roundabout import chain:
service_protocol.py → orchestra.__init__ (lazy) → orchestra.domain_types

Changes:
- heartbeat.py: import from orchestra.domain_types
- runtime/__init__.py: update both TYPE_CHECKING and __getattr__ imports
- test_heartbeat.py: update import
- test_core_budget.py: remove runtime.service_protocol from core modules
- Delete service_protocol.py

Closes #2573
