+++ src/vibe3/runtime/__init__.py
@@ -8,6 +8,7 @@ from typing import TYPE_CHECKING
 
 # Lazy imports via __getattr__ for everything to avoid circular dependencies
 if TYPE_CHECKING:
+    from vibe3.orchestra.domain_types import ServiceBase
     from vibe3.runtime.circuit_breaker import (
         CircuitBreaker,
         CircuitState,
@@ -23,7 +24,6 @@ if TYPE_CHECKING:
         write_instance_info,
     )
     from vibe3.runtime.periodic_check_executor import execute_periodic_check
-    from vibe3.runtime.service_protocol import ServiceBase
     from vibe3.runtime.taxonomy import MODULE_CATEGORY_MAP, ModuleCategory
 
 
@@ -84,7 +84,7 @@ def __getattr__(name: str) -> object:
 
         return execute_periodic_check
     if name == "ServiceBase":
-        from vibe3.runtime.service_protocol import ServiceBase
+        from vibe3.orchestra.domain_types import ServiceBase
 
         return ServiceBase
