--- src/vibe3/runtime/service_protocol.py
@@ -1,10 +1,0 @@
-"""Service protocol for runtime observers.
-
-Re-exports ServiceBase from domain/protocols to maintain backward compatibility.
-Actual definition moved to domain/protocols/runtime_protocols.py to break
-the domain→runtime circular dependency.
-"""
-
-from vibe3.orchestra import ServiceBase
-
-__all__ = ["ServiceBase"]
