+++ tests/vibe3/runtime/test_heartbeat.py
@@ -6,9 +6,9 @@ from unittest.mock import MagicMock
 import pytest
 
 from vibe3.models.orchestra_config import OrchestraConfig
+from vibe3.orchestra.domain_types import ServiceBase
 from vibe3.orchestra.failed_gate import GateResult
 from vibe3.runtime.heartbeat import HeartbeatServer
-from vibe3.runtime.service_protocol import ServiceBase
 
 
 def _config() -> OrchestraConfig:
